+++ user_service/src/main/java/com/digitalbook/user/jwt/UserDetailsImpl.java
@@ -19,7 +19,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 public class UserDetailsImpl implements UserDetails {
 	private static final long serialVersionUID = 1L;
 
-	private int id;
+	private Long id;
 
 	private String username;
 
@@ -30,7 +30,7 @@ public class UserDetailsImpl implements UserDetails {
 
 	private Collection<? extends GrantedAuthority> authorities;
 
-	public UserDetailsImpl(int id, String username, String email, String password,
+	public UserDetailsImpl(Long id, String username, String email, String password,
 			Collection<? extends GrantedAuthority> authorities) {
 		this.id = id;
 		this.username = username;
@@ -56,7 +56,7 @@ public class UserDetailsImpl implements UserDetails {
 		return authorities;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
